% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\docType{data}
\name{cori_colors}
\alias{cori_colors}
\title{Vector of CORI colors}
\format{
An object of class \code{character} of length 13.
}
\usage{
cori_colors
}
\description{
Vector of CORI colors
}
\examples{
cori_colors["Emerald"]

}
\keyword{datasets}
