% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cori_cols.R
\name{cori_cols}
\alias{cori_cols}
\title{Function to extract cori_ctg7 colors as hex codes}
\usage{
cori_cols(...)
}
\arguments{
\item{...}{Character names of cori_ctg7 colors}
}
\description{
Function to extract cori_ctg7 colors as hex codes
}
\examples{
cori_cols("Emerald")

}
