% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cori_pal.R
\name{cori_pal}
\alias{cori_pal}
\title{Return function to interpolate a cori color palette}
\usage{
cori_pal(palette = "ctg2gn", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in cori_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a cori color palette
}
\examples{
cori_pal("ctg7")

}
