% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\docType{data}
\name{cori_palettes}
\alias{cori_palettes}
\title{List of CORI palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
cori_palettes
}
\description{
List of CORI palettes
}
\examples{
cori_palettes[["ctg2gn"]]

}
\keyword{datasets}
