% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ias.r
\docType{data}
\name{ias}
\alias{ias}
\title{Dataset used for creating example plots in the CORI style.}
\format{
Data frame with columns
\describe{
\item{row_id}{row id}
\item{geoid}{}
\item{year}{}
\item{sex}{}
\item{race}{}
\item{percent_high_speed_internet}{}
\item{percent_regular_internet}{}
}
}
\usage{
ias
}
\description{
Dataset used for creating example plots in the CORI style.
}
\keyword{datasets}
