% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbra.R
\docType{data}
\name{pbra}
\alias{pbra}
\title{Dataset used for creating example plots in the CORI style.}
\format{
Data frame with columns
\describe{
\item{row_id}{row id}
\item{geoid}{}
\item{rural_ruca}{}
\item{rural_fhfa}{}
\item{rural_forhp}{}
\item{rural_rucc}{}
\item{rural_uic}{}
\item{rural_cdc}{}
\item{rural_cbsa_2020}{}
\item{rural_cbsa_2013}{}
\item{rural_far}{}
\item{year}{}
\item{population}{}
\item{race}{}
\item{county}{}
}
}
\usage{
pbra
}
\description{
Dataset used for creating example plots in the CORI style.
}
\keyword{datasets}
