% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcu.r
\docType{data}
\name{rcu}
\alias{rcu}
\title{Dataset used for creating example plots in the CORI style.}
\format{
Data frame with columns
\describe{
\item{index}{row id}
\item{st_code}{}
\item{county_code}{}
\item{year}{}
\item{attr}{}
\item{tot_pop}{}
\item{st_cnty}{}
}
}
\usage{
rcu
}
\description{
Dataset used for creating example plots in the CORI style.
}
\keyword{datasets}
