% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsh_pop.r
\docType{data}
\name{rsh_pop}
\alias{rsh_pop}
\title{Dataset used for creating example plots in the CORI style.}
\format{
Data frame with columns
\describe{
\item{geoid}{Row in geofacet}
\item{rural_ruca}{}
\item{rural_fhfa}{}
\item{rural_forhp}{}
\item{rural_rucc}{}
\item{rural_uic}{}
\item{rural_cdc}{}
\item{rural_cbsa_2020}{}
\item{rural_cbsa_2013}{}
\item{rural_far}{}
\item{estimate_agriculture}{}
\item{year}{}
\item{estimate_ba_or_higher}{}
\item{estimate_below_poverty_level}{}
\item{estimate_broadband_accfess}{}
}
}
\usage{
rsh_pop
}
\description{
Dataset used for creating example plots in the CORI style.
}
\keyword{datasets}
