% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_cori.r
\name{scale_fill_cori}
\alias{scale_fill_cori}
\title{Fill scale constructor for cori colors}
\usage{
scale_fill_cori(palette = "ctg2gn", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in cori_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for cori colors
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
geom_point(size = 4, alpha = .6) +
scale_color_cori(discrete = FALSE, palette = "ctg7")

}
