% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cori.r
\name{theme_cori}
\alias{theme_cori}
\title{A \link{ggplot2} theme formatted in the Center on Rural Innovation (CORI) style}
\usage{
theme_cori(base_family = "TT Hoves", base_font_size = 12)
}
\description{
\code{theme_cori} provides \link{ggplot2} theme formatted
according to the CORI style guide.
}
