% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cori_pyramid.R
\name{theme_cori_pyramid}
\alias{theme_cori_pyramid}
\title{A \link{ggplot2} theme formatted in the Center on Rural Innovation (CORI) style}
\usage{
theme_cori_pyramid(base_family = "TT Hoves", base_font_size = 12)
}
\arguments{
\item{base_family, }{the font family that CORI employed, default TT Hoves}

\item{base_font_size, }{the default text font size}
}
\description{
\code{theme_cori_pyramid} provides \link{ggplot2} theme for population pyramids
}
