% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tthoves_install.R
\name{tthoves_install}
\alias{tthoves_install}
\title{Import and register TT Hoves font}
\usage{
tthoves_install()
}
\description{
\code{{tthoves_install}} checks whether TT Hoves is imported and registered.
If the font does not imported and registered, then the \code{{{tthoves_install}}} let you know that the font should be installed the computer. Once, the font installed the local instance, then \code{{tthoves}} will import fonts and register to the R.
}
