% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tthoves_test.r
\name{tthoves_test}
\alias{tthoves_test}
\title{Test whether the TT hoves font are properly imported and registered}
\usage{
tthoves_test()
}
\description{
\code{tthoves_test} tests to see if it is imported and registered.
TT hoves font family is necessary to install
\href{https://www.dropbox.com/sh/s5xw4dbcag403zw/AACekAb3h4GFxBBw0WGeRl0Pa?dl=0}{TT Hoves font}
}
\details{
Import and register TT hoves font in R with tthoves_install()
}
