% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upr.r
\docType{data}
\name{upr}
\alias{upr}
\title{Dataset used for creating example plots in the CORI style.}
\format{
Data frame with columns
\describe{
\item{row_id}{row id}
\item{race}{}
\item{year}{}
\item{population}{}
}
}
\usage{
upr
}
\description{
Dataset used for creating example plots in the CORI style.
}
\keyword{datasets}
