% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cori_legend.r
\name{cori_legend}
\alias{cori_legend}
\title{A \link{ggplot2} legend formatted in the Center on Rural Innovation (CORI) style}
\usage{
cori_legend(title = "", font_size = 14)
}
\arguments{
\item{title}{a title name of legend, default ""}

\item{font_size}{font size of legend, default 14}
}
\description{
\code{cori_legend} provides \link{ggplot2} theme formatted
according to the CORI style guide.
}
